format imports and move sentinel specific functions to separate files
