+++ sentle/reproject_util.py
@@ -1,6 +1,7 @@
 import warnings
+
 from affine import Affine
-from rasterio import warp, windows
+from rasterio import transform, warp, windows
 
 
 def check_and_round_bounds(left, bottom, right, top, res):
@@ -126,3 +127,26 @@ def recrop_write_window(win, overall_height, overall_width):
         width=gwidth).round_offsets().round_lengths(), windows.Window(
             row_off=lrow, col_off=lcol, height=lheight,
             width=lwidth).round_offsets().round_lengths()
+
+
+def bounds_from_transform_height_width_res(tf, height, width, resolution):
+    # minx, miny, maxx, maxy
+    return (tf.c, tf.f - (height * resolution), tf.c + (width * resolution),
+            tf.f)
+
+
+def transform_height_width_from_bounds_res(left, bottom, right, top, res):
+    width, rem = divmod(right - left, res)
+    assert rem == 0
+    width = int(width)
+    height, rem = divmod(top - bottom, res)
+    assert rem == 0
+    height = int(height)
+    tf = transform.from_bounds(west=left,
+                               south=bottom,
+                               east=right,
+                               north=top,
+                               width=width,
+                               height=height)
+
+    return tf, height, width
