+++ sentle/stac.py
@@ -1,9 +1,9 @@
-from .const import STAC_ENDPOINT
-from pystac_client.stac_api_io import StacApiIO
 import planetary_computer
 import pystac_client
-
+from pystac_client.stac_api_io import StacApiIO
 from urllib3 import Retry
+
+from .const import STAC_ENDPOINT
 
 
 def get_stac_api_io():
