+++ sentle/utils.py
@@ -1,28 +1,7 @@
-from rasterio import transform
 import contextlib
+
 import joblib
 
-def bounds_from_transform_height_width_res(tf, height, width, resolution):
-    # minx, miny, maxx, maxy
-    return (tf.c, tf.f - (height * resolution), tf.c + (width * resolution),
-            tf.f)
-
-
-def transform_height_width_from_bounds_res(left, bottom, right, top, res):
-    width, rem = divmod(right - left, res)
-    assert rem == 0
-    width = int(width)
-    height, rem = divmod(top - bottom, res)
-    assert rem == 0
-    height = int(height)
-    tf = transform.from_bounds(west=left,
-                               south=bottom,
-                               east=right,
-                               north=top,
-                               width=width,
-                               height=height)
-
-    return tf, height, width
 
 @contextlib.contextmanager
 def tqdm_joblib(tqdm_object):
@@ -46,4 +25,3 @@ def tqdm_joblib(tqdm_object):
     finally:
         joblib.parallel.BatchCompletionCallBack = old_batch_callback
         tqdm_object.close()
-
